% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinnishSentiment.R
\docType{package}
\name{FinnishSentiment}
\alias{FinnishSentiment}
\title{FinnishSentiment: Sentiment Analysis for Finnish Language.}
\description{
Machine learning models and code for sentiment analysis in Finnish.
}
